refactor: move unit tests from src to tests directories
Move 152 tests from inline #[cfg(test)] modules to dedicated tests/
directories while keeping tests that require internal access inline.

Changes by crate:
- docspec-core: 135 tests to tests/*.rs, 17 stay inline in src/stack.rs
- docspec-blocknote-writer: 6 tests to tests/writer.rs
- docspec-wasm: 4 tests to tests/wasm.rs
- docspec-cli: 14 tests stay inline (private modules)

Also fixes 5 silent test failures in writer.rs by converting
if-let-ok patterns to explicit .expect() assertions.

Refs: #13
